feat(markdown-content): Create pokemon service to use it in ContentController and PokemonController and don't repeat the code
